on_prem_llm_generation_job/requirements.txt
fastapi
uvicorn
websockets
llama-cpp-python[server]
requests